[MIN-153] Remove unpauseSpecificOperation.
