+++ src/actions/economicData.ts
@@ -1,4 +1,3 @@
-import { web3FromAddress } from '@polkadot/extension-dapp';
 import { Dispatch } from '../util/types';
 import {
 	RESET_ADMIN_REQUESTS,
@@ -14,13 +13,9 @@ import {
 	GET_PAUSE_KEEPERS_START,
 	GET_PAUSE_KEEPERS_SUCCESS,
 	GET_PAUSE_KEEPERS_ERROR,
-	UNPAUSE_SPECIFIC_OPERATION_START,
-	UNPAUSE_SPECIFIC_OPERATION_SUCCESS,
-	UNPAUSE_SPECIFIC_OPERATION_ERROR,
 } from './types';
 import API from '../services';
 import { UNDERLYING_ASSETS_TYPES } from '../util/constants';
-import { txCallback } from '../util';
 
 export const resetAdminRequests = () => {
 	return {
@@ -140,40 +135,3 @@ export const getPauseKeepers = () => {
 		}
 	};
 };
-
-export function unpauseSpecificOperation(
-	account: string,
-	keyring: any,
-	poolId: string,
-	operation: string
-) {
-	return async (dispatch: Dispatch) => {
-		const callBack = txCallback(
-			[UNPAUSE_SPECIFIC_OPERATION_SUCCESS, UNPAUSE_SPECIFIC_OPERATION_ERROR],
-			dispatch
-		);
-
-		try {
-			dispatch({ type: UNPAUSE_SPECIFIC_OPERATION_START });
-			const currentUser = keyring.getPair(account);
-
-			if (currentUser.isLocked) {
-				const injector = await web3FromAddress(account);
-				await API.tx.sudo
-					.sudo(API.tx.controller.unpauseSpecificOperation(poolId, operation))
-					// @ts-ignore
-					.signAndSend(account, { signer: injector.signer }, callBack);
-			} else {
-				await API.tx.sudo
-					.sudo(API.tx.controller.unpauseSpecificOperation(poolId, operation))
-					// @ts-ignore
-					.signAndSend(currentUser, callBack);
-			}
-		} catch (err) {
-			dispatch({
-				type: UNPAUSE_SPECIFIC_OPERATION_ERROR,
-				payload: err.toString(),
-			});
-		}
-	};
-}
